rename conversionview to displaybar

+++ imagewindow.h
@@ -11,7 +11,7 @@ void image_window_open( ImageWindow *win, GFile *file );
 double image_window_get_scale( ImageWindow *win );
 TileSource *image_window_get_tilesource( ImageWindow *win );
 void image_window_set_tile_source( ImageWindow *win, TileSource *tile_source );
-void image_window_get_last( ImageWindow *win, int *last_x, int *last_y );
+void image_window_get_last( ImageWindow *win, double *last_x, double *last_y );
 
 #endif /* __IMAGE_WINDOW_H */
